implementing UI

+++ routes/web.php
@@ -1,18 +1,27 @@
 <?php
 
-Route::get('/', 'Auth\LoginController@loginRedirect');
+Route::get('/admin/login', 'Auth\LoginController@loginRedirect');
 
-Route::group(['middleware' => ['auth']], function() {
-    Route::get('/home', 'HomeController@index');
-    Route::resource('roles','RoleController');
-    Route::resource('users','UserController');
-    Route::resource('articles','ArticleController');
-    Route::resource('galery','GaleryController');
-    Route::resource('contact','ContactController', ['only' => ['index', 'update']]);
-    Route::resource('profil','ProfilController');
-    Route::resource('permission','PermissionController');
-    Route::delete('/roles/permission/{permission}', 'RoleController@revokePermission');
-    Route::post('/roles/permission', 'RoleController@assignPermission');
+Route::get('/', 'HomeController@index');
+Route::get('/car', 'CarController@index');
+Route::get('/car/detail', 'CarController@detail');
+
+Route::get('/galery', 'GaleryController@index');
+Route::get('/galery/detail', 'GaleryController@detail');
+
+Route::get('/contact', 'ContactController@index');
+
+Route::group(['middleware' => ['auth'], 'prefix'=>'admin'], function() {
+    Route::get('/home', 'admin\HomeController@index');
+    Route::resource('roles','admin\RoleController');
+    Route::resource('users','admin\UserController');
+    Route::resource('articles','admin\ArticleController');
+    Route::resource('galery','admin\GaleryController');
+    Route::resource('contact','admin\ContactController', ['only' => ['index', 'update']]);
+    Route::resource('profil','admin\ProfilController');
+    Route::resource('permission','admin\PermissionController');
+    Route::delete('/roles/permission/{permission}', 'admin\RoleController@revokePermission');
+    Route::post('/roles/permission', 'admin\RoleController@assignPermission');
 });
 
 Auth::routes();
